Nia's Pieces Added
The flow generally goes:

1. SheronySp1... houses the main method which
2. Opens the InputsCntl (serving as the controller) that creates several
Runs (model object) which should be created by Dana's part in the real
project. Then
3. The interface (ScannerInputs) is opened by the controller. Within the
inputs, the code for the submit button  checks to see if it's the last
"run" in the runlist If not, it saves the info the user inputs as a new
run and opens a new, identical dialog for the next "run".
4. If it is the last run, the view triggers the print method in the
controller.